# build: pkg cli in npm