--- a/.circleci/config.base.yml
+++ b/.circleci/config.base.yml
@@ -198,6 +198,7 @@ jobs:
             yarn yarn-use-bash
             yarn pkg-all
             unsetNpmRegistryUrl
+            uploadPkgCli
       - save_cache:
           key: amplify-pkg-binaries-{{ .Branch }}-{{ .Revision }}
           paths:
@@ -243,30 +244,6 @@ jobs:
             sudo npm install -g @aws-amplify/cli
             unsetSudoNpmRegistryUrl
             amplify version
-
-  amplify_e2e_tests:
-    <<: *defaults
-    steps:
-      - attach_workspace:
-          at: ./
-      - restore_cache:
-          key: amplify-cli-yarn-deps-{{ .Branch }}-{{ checksum "yarn.lock" }}-{{ arch }}
-      - restore_cache:
-          key: amplify-verdaccio-cache-{{ .Branch }}-{{ .Revision }}
-      - restore_cache:
-          key: amplify-pkg-binaries-{{ .Branch }}-{{ .Revision }}
-      - run: *install_cli_from_local_registery
-      - run:
-          name: Run E2e Tests
-          command: |
-            source .circleci/local_publish_helpers.sh
-            retry runE2eTest
-          no_output_timeout: 90m
-      - run: *scan_e2e_test_artifacts
-      - store_test_results:
-          path: packages/amplify-e2e-tests/
-      - store_artifacts:
-          path: ~/repo/packages/amplify-e2e-tests/amplify-e2e-reports
 
   amplify_e2e_tests_pkg:
     parameters:
@@ -543,6 +520,32 @@ jobs:
           keys:
             - amplify-cli-ssh-deps-{{ .Branch }}
       - run:
+          name: Upload Pkg Binary
+          command: |
+            sudo apt-get install -y lsof
+            git config user.email not@used.com
+            git config user.name "Doesnt Matter"
+            source .circleci/local_publish_helpers.sh
+            yarn yarn-use-bash
+            yarn verdaccio-clean
+            yarn verdaccio-disconnect
+            yarn verdaccio-stop
+            startLocalRegistry "$(pwd)/.circleci/verdaccio.yaml"
+            setNpmRegistryUrlToLocal
+            changeNpmGlobalPath
+            setNpmTag
+            if [ -z $NPM_TAG ]; then
+              yarn publish-to-verdaccio
+            else
+              yarn lerna publish --exact --dist-tag=latest --preid=$NPM_TAG --conventional-commits --conventional-prerelease --no-verify-access --yes --no-commit-hooks --no-push --no-git-tag-version
+            fi
+            yarn pkg-all
+            unsetNpmRegistryUrl
+            git stash
+            git stash drop
+            uploadPkgCli
+            ./out/amplify-pkg-linux --version
+      - run:
           name: Authenticate with npm
           command: echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc
       - run:
@@ -654,6 +657,13 @@ workflows:
       - publish_to_local_registry:
           requires:
             - build
+      - build_pkg_binaries:
+          context:
+            - e2e-auth-credentials
+            - e2e-test-context
+            - amplify-s3-upload
+          requires:
+            - publish_to_local_registry
       - amplify_console_integration_tests:
           context:
             - amplify-ecr-image-pull
@@ -661,8 +671,7 @@ workflows:
             - e2e-auth-credentials
             - e2e-test-context
           requires:
-            - build
-            - publish_to_local_registry
+            - build_pkg_binaries
   e2e_resource_cleanup:
     triggers:
       - schedule:
@@ -708,7 +717,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - build
+            - build_pkg_binaries
       - integration_test:
           context:
             - amplify-ecr-image-pull
@@ -734,34 +743,22 @@ workflows:
           requires:
             - build
       - build_pkg_binaries:
+          context:
+            - e2e-auth-credentials
+            - e2e-test-context
+            - amplify-s3-upload
           requires:
             - publish_to_local_registry
       - amplify_sudo_install_test:
           context: amplify-ecr-image-pull
           requires:
-            - publish_to_local_registry
+            - build_pkg_binaries
           filters:
             branches:
               only:
                 - master
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
-      - amplify_e2e_tests:
-          context:
-            - amplify-ecr-image-pull
-            - cleanup-resources
-            - e2e-auth-credentials
-            - e2e-test-context
-          # post-steps:
-          #   - run: *clean_e2e_resources
-          filters:
-            branches:
-              only:
-                - master
-                - /tagged-release\/.*/
-                - /run-e2e\/.*/
-          requires:
-            - publish_to_local_registry
       - amplify_e2e_tests_pkg:
           context:
             - amplify-ecr-image-pull
@@ -789,7 +786,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - build
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_v5:
@@ -805,7 +802,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - build
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_non_multi_env_layers:
@@ -821,7 +818,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - build
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_multi_env_layers:
@@ -837,7 +834,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - build
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_console_integration_tests:
@@ -854,8 +851,7 @@ workflows:
               only:
                 - beta
           requires:
-            - build
-            - publish_to_local_registry
+            - build_pkg_binaries
       - github_prerelease:
           context: github-publish
           requires:
@@ -876,7 +872,6 @@ workflows:
             - cleanup-resources
             - e2e-test-context
           requires:
-            - amplify_e2e_tests
             - amplify_e2e_tests_pkg
             - amplify_migration_tests_v6
             - amplify_migration_tests_v5
@@ -884,13 +879,15 @@ workflows:
       - deploy:
           context:
             - amplify-ecr-image-pull
+            - e2e-auth-credentials
+            - e2e-test-context
+            - amplify-s3-upload
             - npm-publish
           requires:
             - test
             - mock_e2e_tests
             - graphql_e2e_tests
             - integration_test
-            - amplify_e2e_tests
             - amplify_e2e_tests_pkg
             - amplify_sudo_install_test
             - amplify_console_integration_tests