--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -214,6 +214,7 @@ module.exports = {
     '/packages/amplify-category-function/lib',
     '/packages/amplify-category-geo/lib',
     '/packages/amplify-category-storage/lib',
+    '/packages/amplify-cli-npm/lib',
     '/packages/amplify-cli-core/lib',
     '/packages/amplify-cli/lib',
     '/packages/amplify-cli-logger/lib',