--- a/.gitignore
+++ b/.gitignore
@@ -28,6 +28,7 @@ packages/amplify-dynamodb-simulator/__test__/dynamodb-data
 packages/graphql*/lib
 packages/amplify-graphql-transformer-*/lib
 packages/amplify-cli/lib
+packages/amplify-cli-npm/lib
 packages/amplify-cli-core/lib
 packages/amplify-cli-logger/lib
 packages/amplify-mqtt-server/lib