--- a/packages/amplify-app/src/scripts/amplify-modelgen.js
+++ b/packages/amplify-app/src/scripts/amplify-modelgen.js
@@ -7,7 +7,7 @@ console.log('Running codegen...');
 run();
 
 async function run() {
-  const amplify = /^win/.test(process.platform) ? 'amplify.cmd' : 'amplify';
+  const amplify = process.env.AMPLIFY_PATH ? process.env.AMPLIFY_PATH : /^win/.test(process.platform) ? 'amplify.cmd' : 'amplify';
   const modelGen = spawn(amplify, ['codegen', 'model'], { cwd: process.cwd(), env: process.env, stdio: 'inherit' });
 
   modelGen.on('exit', code => {