--- a/packages/amplify-cli-npm/.npmignore
+++ b/packages/amplify-cli-npm/.npmignore
@@ -1,0 +1,2 @@
+bin/*
+tsconfig.tsbuildinfo