--- a/packages/amplify-cli-npm/install.ts
+++ b/packages/amplify-cli-npm/install.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { install } from '.';
+
+install();