--- a/packages/amplify-cli-npm/package.json
+++ b/packages/amplify-cli-npm/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@aws-amplify/cli",
+  "version": "7.6.26",
+  "description": "Amplify CLI",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/aws-amplify/amplify-cli.git",
+    "directory": "packages/amplify-cli"
+  },
+  "author": "Amazon Web Services",
+  "license": "Apache-2.0",
+  "keywords": [
+    "graphql",
+    "appsync",
+    "aws"
+  ],
+  "bin": {
+    "amplify": "./lib/run.js"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "postinstall": "node ./lib/install.js || echo \"failed to install amplify binary\"",
+    "prepublishOnly": "rimraf ./lib && tsc",
+    "run": "ts-node ./run.ts",
+    "build": "tsc"
+  },
+  "engines": {
+    "node": ">=12.0.0"
+  },
+  "dependencies": {
+    "axios": "^0.26.0",
+    "rimraf": "^3.0.2"
+  },
+  "devDependencies": {
+    "prettier": "^1.19.1",
+    "rimraf": "^3.0.2"
+  }
+}