--- a/packages/amplify-cli-npm/run.ts
+++ b/packages/amplify-cli-npm/run.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { run } from '.';
+
+run();