--- a/packages/amplify-e2e-tests/src/__tests__/containers-api-secrets.test.ts
+++ b/packages/amplify-e2e-tests/src/__tests__/containers-api-secrets.test.ts
@@ -6,11 +6,12 @@ import {
   deleteProject,
   deleteProjectDir,
   initJSProjectWithProfile,
+  retry,
 } from 'amplify-e2e-core';
 import fetch from 'node-fetch';
-import { getAWSExports } from '../aws-exports/awsExports';
 import * as fs from 'fs-extra';
 import path from 'path';
+import { getAWSExports } from '../aws-exports/awsExports';
 
 async function setupAmplifyProject(cwd: string) {
   await amplifyConfigureProject({
@@ -36,7 +37,7 @@ describe('amplify api add', () => {
     await initJSProjectWithProfile(projRoot, { name: 'multicontainer', envName });
     await setupAmplifyProject(projRoot);
     await addRestContainerApi(projRoot, { apiName });
-    await setupContainerSecrets(projRoot, apiName)
+    await setupContainerSecrets(projRoot, apiName);
     await amplifyPushSecretsWithoutCodegen(projRoot);
     const awsExports: any = getAWSExports(projRoot).default;
     const {
@@ -44,11 +45,21 @@ describe('amplify api add', () => {
     } = awsExports;
     expect(name).toBeDefined();
     expect(endpoint).toBeDefined();
-
-    const result = await (await fetch(`${endpoint}/password`)).text();
-    expect(result).toEqual('CONTAINER_SECRETS_PASSWORD');
+    const url = `${endpoint}/password`;
+    const expected = 'CONTAINER_SECRETS_PASSWORD';
+    const result = await retry(
+      async (): Promise<string> => (await fetch(url)).text(),
+      (fetchResult: string) => fetchResult === expected,
+      {
+        times: 100,
+        delayMS: 100,
+        // five minutes
+        timeoutMS: 300000,
+        stopOnError: false,
+      },
+    );
+    expect(result).toEqual(expected);
   });
-
 });
 
 const setupContainerSecrets = async (projRoot: string, apiName: string) => {
@@ -59,7 +70,7 @@ const setupContainerSecrets = async (projRoot: string, apiName: string) => {
   const secretsFolder = path.join(__dirname, '..', '..', 'resources', 'api-container', 'secrets');
   const dockerFile = path.join(__dirname, '..', '..', 'resources', 'api-container', 'docker-compose.yml');
   const expressFile = path.join(__dirname, '..', '..', 'resources', 'api-container', 'express', 'secret-password-index.js');
-  
+
   // Read the docker and express index files from resources
   const dockerFileContents = await fs.readFile(dockerFile);
   const expressFileContents = await fs.readFile(expressFile);