Due to the size of this changeset, I recommend reviewing this pull request using [VS Code's GitHub pull request review plugin](https://code.visualstudio.com/blogs/2018/09/10/introducing-github-pullrequests). The changeset becomes much smaller when you can visualize that the majority of the changed files are moved/renamed.

Most significant changes:
* packages/amplify-cli has been renamed to packages/amplify-cli-builder. This is to accommodate maintaining all existing behavior while deploying entirely new source code for `@aws-amplify/cli`.
* packages/amplify-cli has been replaced with a small code package that downloads the pkg-cli binary from CloudFront as a post install script and routes all commands to the downloaded binary.
* .circleci/config.base.yml has been updated to upload pkg cli versions during test/deploy phases. Tests of the non-pkg cli have been  removed (with exceptions being `amplify_sudo_install_test` and migration tests.)

Tagged release is available at `npm install -g @aws-amplify/cli@7.6.26-npm-pkg-cli.0`

Benefits: 
- pkg cli decreases installation time by 92.63% from 216.81s to 15.974s.
- pkg cli enjoys a 32.9% runtime performance benefit for example on the `amplify --version` command, from 1.608s to 1.079s.
- pkg cli decreases storage use on client machines by 35.71% from 924M to 594M.
- pkg cli guarantees parity across all users for a specific version regardless of consumption mechanism (npm vs raw pkg cli download) 
- redundant e2e tests are removed, reducing per-run CCI credit usage by ~40%.
- pkg cli guarantees dependencies are tested, preventing supply chain attacks a la `ua-parser-js` and `colors`.

See also [tagged release test CCI job](https://app.circleci.com/pipelines/github/aws-amplify/amplify-cli?branch=tagged-release-without-e2e-tests/npm-pkg-cli-test)
See also [full e2e suite using pkg cli](https://app.circleci.com/pipelines/github/aws-amplify/amplify-cli?branch=run-e2e/xss/npm-pkg-cli-test)